Fix guest_multiplier saving 0.5 as 0 (parseInt -> parseFloat)
parseInt strips the decimal so 0.5 was saved as 0. parseFloat handles
fractional values correctly.

+++ app/(app)/products/finished/page.tsx
@@ -126,7 +126,7 @@ export default function FinishedGoodsPage() {
       vat_rate:         editState.vat === '19' ? 0.19 : 0.07,
       occasion:         editState.occasion,
       menu_category:    editState.menu_category,
-      guest_multiplier: parseInt(editState.guest_multiplier)   || 0,
+      guest_multiplier: parseFloat(editState.guest_multiplier)  || 0,
     }).eq('id', id);
     setSaving(false);
     if (!error) {
